+++ packages/cloud_function/src/__tests__/funcs/test.func.ts
@@ -1,0 +1,7 @@
+import { useFunc } from '@faasjs/func'
+
+export const func = useFunc(() => {
+  return async (event) => {
+    return event
+  }
+})